delete calling of update_env_with_obj_type_info
update_env_with_obj_type_info was used to add base object fields (when
type of base object was inferred before inheritance) to the new type env
which was created to infer type of Inherit node. Since it was temporary
solution and now the concept of required flags is introduced, we don't
need this function anymore.