add logging to see the work of hm_algo
Adding logs caused failures in some unit tests where expected result
contains type variables. The reason is in the specific implementation of
class TypeVariable and how it is printed. So, when other variables are
printed within program, then first letters for variable names are
reserved. Then, when we previously type variable name was 'a', now it
can be 'j'.