change row unification and analyse() for LetrecAnd
Update row b with fields of row a if those fields are not optional and
are not in row b.
Replace defn_type and v_type in unify call such that now defn_type
corresponds to b and v_type to a. Then, add unified defn_type to
new_env.
Some unit tests are failing because optional flags are not added to the
fields so far and order of fields is changed. Unit tests will be fixed
when some other necessary parts of row polymorphism will be implemented.